Improve performance of stack trace calls
In `dape--update' request one frame of stack trace, then in info
buffer request the rest based on new custom `dape-stack-trace-levels'.
This will result in a double render of when adapter
:supportsDelayedStackTraceLoading, but when the adapter times out on
the stack-trace request this is an overall improvement.

This change came about due to an issue where js-debug stackTrace would
timeout which it still does if it's an expensive one,  "WASM" and
stuff in the stack. But state is not ruind only the info buffer is
under populated.

Why does this not fail on vscode? seams like vscode uses some kind of
caching together with js-debug. As the stackTrace call will not take
the same amount of time whenever dap visits the part of code again, so
I am guessing vscode supplies some cache part or some other hidden
"launch" request parameter.