Replace core dape macros with hygienic macros
- Remove dape--callback and dape--with
- Add dape--with-request, dape--with-request-bind macros for clarity

This change was not done not to make the macros hygienic but to reduce
the barrier to entry for other contributes then me.

- dape--with-request does not bind any special vars like dape--with
- dape--with-request-bind works and looks like cl-destructuring-bind
- dape--request-return to "standardize" the request callback function
  signature

If you lift the carpet, you will still find the same callback design
which might scare people away but it's my opinion that it at least
looks like some cl influenced piece of emacs lisp code.