Move `dape-on-stopped-hooks' back to it's original place
There is nothing in the manual that says that a thread needs to be
stopped so let's just use the last adapter for reading and writing
memory.  If that is the case we should revert the hook call point as
it makes more sense at the end of stopped event.